xdiff: make xdfile_t.nrec a size_t instead of long
size_t is used because nrec describes the number of elements in memory
for recs, and the number of elements in memory for 'changed' + 2.

+++ xdiff/xtypes.h
@@ -47,7 +47,7 @@ typedef struct s_xrecord {
 
 typedef struct s_xdfile {
 	xrecord_t *recs;
-	long nrec;
+	size_t nrec;
 	bool *changed;
 	long *rindex;
 	long nreff;
